+++ lib/lotus/storage/query.ex
@@ -14,10 +14,10 @@ defmodule Lotus.Storage.Query do
   require Logger
 
   alias Lotus.Config
+  alias Lotus.Query.OptionalClause
   alias Lotus.Query.Statement
   alias Lotus.Source
   alias Lotus.Source.Adapter
-  alias Lotus.Query.OptionalClause
   alias Lotus.Storage.{QueryVariable, SchemaCache, TypeCaster, VariableResolver}
 
   @type t :: %__MODULE__{